+++ packages/landing/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.13.0](https://github.com/linz/basemaps/compare/v4.12.2...v4.13.0) (2020-09-14)
+
+**Note:** Version bump only for package @basemaps/landing
+
+
+
+
+
+## [4.12.2](https://github.com/linz/basemaps/compare/v4.12.0...v4.12.2) (2020-09-10)
+
+**Note:** Version bump only for package @basemaps/landing
+
+
+
+
+
+## [4.12.1](https://github.com/linz/basemaps/compare/v4.12.0...v4.12.1) (2020-09-10)
+
+**Note:** Version bump only for package @basemaps/landing
+
+
+
+
+
 # [4.12.0](https://github.com/linz/basemaps/compare/v4.11.2...v4.12.0) (2020-09-06)